feat: checkout url on booking details
